Sync Overhaul Part 1: Removed a bunch of avoidable resyncs and fixed a bunch of syncing bugs.
Renamed syncToGame to sync and added a syncSilent function that is called for the initial sync.
No longer generate a matchError if the only error is that the centuryMultiplier has changed. Instead
 - the Simulator.matchError function will quickly fix this and recheck before moving on.
No longer generate a matchError when the Buffs change. Buffs are automatically resynced whenever the
 - number of buffs in the game changes.
No longer generate a matchError when you buy a building while a scaling upgrade of that type is
 - active. The cached scaling factor is now used when it is set.
No longer log the initial sync.

Refactored the following:
 - Updated all remaining modifier functions to use () => syntax and no longer take sim as a parameter.
 - Added type information to function parameters in the Buff class.